% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{attr_path}
\alias{attr_path}
\title{Attribute Paths}
\usage{
attr_path(path, dates, value, outcome, lookup)
}
\arguments{
\item{path}{Character vector. Channels delimited by ">"}

\item{dates}{Character vector. Dates delimeted by ">"}

\item{value}{Numeric vector.}

\item{outcome}{Numeric vector.}

\item{lookup}{a named list of lookup values. Each element must be a scalar numeric value corresponding to a path value.}
}
\description{
Attribute Paths
}
