% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original-pathattr.R
\name{attribute_path}
\alias{attribute_path}
\title{function to convert a single path to conversion to a data frame with
one record for each touchpoint showing the value of that touchpoint}
\usage{
attribute_path(path_str, date_str, outcome, value, retbl)
}
\arguments{
\item{path_str}{string of touchpoints like "social>organic>paid_search"
of interest.}

\item{date_str}{string of dates of touchpoints like  "2023-01-01>2023-01-05>2023-01-06"}

\item{outcome}{numeric value, number of successes for the path of touchpoints, like 1}

\item{value}{numeric value, intended to be profit for the outcomes, like 2061}

\item{retbl}{data.frame-a-like containing the removal effects of all touchpoints}
}
\description{
function to convert a single path to conversion to a data frame with
one record for each touchpoint showing the value of that touchpoint
}
